requirements.txt
langchain==0.3.20
python-dotenv==1.0.1
streamlit==1.43.2
unstructured==0.17.0
tiktoken==0.9.0
faiss-cpu==1.10.0
libmagic==1.0
python-libmagic==0.4.0
python-magic-bin==0.4.14
openai==1.66.3
python-magic==0.4.27







langchain-community~=0.3.19